#7 Containerised Flask App with Docker
Reorganised directory structure to follow better conventions
