--- Procfile
@@ -1,1 +1,0 @@
-web: python app.py
