+++ app/__init__.py
@@ -1,0 +1,12 @@
+from flask import Flask
+
+
+def create_app():
+    app = Flask(__name__)
+
+    # Register blueprints
+    from .main import main as main_blueprint
+
+    app.register_blueprint(main_blueprint)
+
+    return app
