+++ scraper/preprocessor.py
@@ -1,0 +1,7 @@
+
+class DataProcessor:
+    def __init__(self, config):
+        self.config = config
+
+    def process(self, data):
+
